# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "T-Rader API",
    "description": "日本株投資プラットフォーム T-Rader の API",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/strategies/{id}/chat": {
      "post": {
        "tags": [
          "strategies"
        ],
        "summary": "フローティングチャットから戦略 Agent にタスクを投入する",
        "operationId": "submit_strategy_chat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "戦略 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "prompt が空 (空白のみを含む)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "戦略が存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type ヘッダが application/json ではない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "リクエストボディのパースに失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "agent task client が未設定",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/strategies/{id}/tasks": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "戦略の過去タスクを新しい順に一覧取得する",
        "operationId": "list_strategy_tasks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "戦略 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyTaskSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "パスパラメータが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "戦略が存在しない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{id}/tasks/{task_id}": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "投入済み戦略タスクの phase / error_summary を取得する",
        "operationId": "get_strategy_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "戦略 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "戦略タスク ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyTaskStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "パスパラメータが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "API エラーレスポンスの JSON 構造",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "エラーメッセージ"
          }
        }
      },
...
      "StrategyChatRequest": {
        "type": "object",
        "description": "フローティングチャットから戦略 Agent に投入する 1 メッセージ。",
        "required": [
          "prompt"
        ],
        "properties": {
          "prompt": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "StrategyChatResponse": {
        "type": "object",
        "description": "`POST /api/strategies/:id/chat` の戻り値。後続の polling 用 task 識別子を返す。",
        "required": [
          "task_id",
          "a2a_task_id"
        ],
        "properties": {
          "a2a_task_id": {
            "type": "string"
          },
          "task_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "StrategyTaskStatusResponse": {
        "type": "object",
        "description": "`GET /api/strategies/:id/tasks/:task_id` の戻り値。",
        "required": [
          "task_id",
          "strategy_id",
          "source",
          "prompt",
          "phase",
          "created_at",
          "updated_at",
          "steps"
        ],
        "properties": {
          "a2a_task_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "result_text": {
            "type": [
              "string",
              "null"
            ],
            "description": "agent の最終応答テキスト (completed 時のみ)"
          },
          "source": {
            "type": "string"
          },
          "steps": {
            "description": "フェーズ/分岐ごとの実行状況。中身は解釈せず素通しする。"
          },
          "strategy_id": {
            "type": "string",
            "format": "uuid"
          },
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StrategyTaskSummary": {
        "type": "object",
        "description": "`GET /api/strategies/:id/tasks` の一覧要素。`steps`/`result_text` は一覧では\n使わないため含めない。",
        "required": [
          "task_id",
          "source",
          "prompt",
          "phase",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}